requirements.txt
flask
mysql-connector-python
yt-dlp
